revert blender utils to isolate fixes
